requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
PyMuPDF==1.24.3
python-docx==1.1.2
Pillow==10.4.0
google-genai
requests==2.31.0
python-dotenv==1.0.1